requirements.txt
datasets[audio]
transformers
accelerate
evaluate
jiwer
tensorboard
gradio
python-dotenv
peft
mlflow